+++ src/svm/test_svm_fifo1.c
@@ -273,7 +273,7 @@ test_ssvm_fifo1 (unformat_input_t * input)
   int verbose = 0;
   int test_id = 0;
 
-  svm_fifo_segment_main_init (0x200000000ULL, 20);
+  svm_fifo_segment_main_init ((ssize_t)0x200000000ULL, 20);
 
   while (unformat_check_input (input) != UNFORMAT_END_OF_INPUT)
     {
